+++ pkg/device/device_service/get_device.go
@@ -6,6 +6,6 @@ import (
 	"data_monitor/pkg/device/device_model"
 )
 
-func GetDevice(id int) (device *device_model.Device, err error) {
+func GetDevice(id int64) (device *device_model.Device, err error) {
 	return device_dao.GetDevice(dao.GetDB(), id)
 }
